+++ server/src/test/java/org/elasticsearch/search/query/QueryPhaseCollectorTests.java
@@ -19,6 +19,7 @@ import org.apache.lucene.search.BooleanQuery;
 import org.apache.lucene.search.BoostQuery;
 import org.apache.lucene.search.CollectionTerminatedException;
 import org.apache.lucene.search.Collector;
+import org.apache.lucene.search.CollectorManager;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.search.FilterCollector;
 import org.apache.lucene.search.FilterLeafCollector;
@@ -42,23 +43,25 @@ import org.elasticsearch.core.IOUtils;
 import org.elasticsearch.test.ESTestCase;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.Matchers;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
 
 public class QueryPhaseCollectorTests extends ESTestCase {
+    private static Directory directory;
+    private static IndexReader reader;
+    private static IndexSearcher searcher;
+    private static int numDocs;
+    private static int numField2Docs;
+    private static int numField3Docs;
+    private static int numField2AndField3Docs;
 
-    private Directory directory;
-    private IndexReader reader;
-    private IndexSearcher searcher;
-    private int numDocs;
-    private int numField2Docs;
-    private int numField3Docs;
-    private int numField2AndField3Docs;
-
-    @Override
-    public void setUp() throws Exception {
-        super.setUp();
+    @BeforeClass
+    public static void beforeClass() throws Exception {
         directory = newDirectory();
         RandomIndexWriter writer = new RandomIndexWriter(random(), directory, newIndexWriterConfig());
         numDocs = randomIntBetween(900, 1000);
@@ -86,10 +89,12 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         writer.close();
     }
 
-    @Override
-    public void tearDown() throws Exception {
-        super.tearDown();
+    @AfterClass
+    public static void afterClass() throws Exception {
         IOUtils.close(reader, directory);
+        searcher = null;
+        reader = null;
+        directory = null;
     }
 
     public void testNullTopDocsCollector() {
@@ -103,88 +108,166 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         );
     }
 
+    /**
+     * Collector manager used temporarily to bridge tests and query phase as they have different expectations when it comes to their
+     * return type and how results are retrieved from a collector manager.
+     */
+    private static class CollectorManagerAdapter<C extends Collector, T> implements CollectorManager<Collector, Void> {
+        private final CollectorManager<C, T> wrapped;
+        private T result;
+
+        CollectorManagerAdapter(CollectorManager<C, T> wrapped) {
+            this.wrapped = wrapped;
+        }
+
+        @Override
+        public Collector newCollector() throws IOException {
+            return wrapped.newCollector();
+        }
+
+        @Override
+        public Void reduce(Collection<Collector> collectors) throws IOException {
+            @SuppressWarnings("unchecked")
+            List<C> cs = collectors.stream().map(collector -> (C) collector).toList();
+            result = wrapped.reduce(cs);
+            return null;
+        }
+
+        public T getResult() {
+            return result;
+        }
+    }
+
     public void testTopDocsOnly() throws IOException {
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, null, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topScoreDocManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topScoreDocManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(topScoreDocAdapter, null, 0, null, null);
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topScoreDocAdapter.getResult().totalHits.value);
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, null, null);
-            searcher.search(new TermQuery(new Term("field2", "value")), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topScoreDocManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topScoreDocManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(topScoreDocAdapter, null, 0, null, null);
+            searcher.search(new TermQuery(new Term("field2", "value")), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
         }
     }
 
     public void testWithAggs() throws IOException {
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, aggsCollector, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(numDocs, aggsCollector.getTotalHits());
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                null,
+                0,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(numDocs, aggsAdapter.getResult().intValue());
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, aggsCollector, null);
-            searcher.search(new TermQuery(new Term("field2", "value")), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(numField2Docs, aggsCollector.getTotalHits());
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                null,
+                0,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new TermQuery(new Term("field2", "value")), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(numField2Docs, aggsAdapter.getResult().intValue());
         }
     }
 
     public void testPostFilterTopDocsOnly() throws IOException {
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
             TermQuery termQuery = new TermQuery(new Term("field2", "value"));
             Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, null, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                null,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
             TermQuery termQuery = new TermQuery(new Term("field1", "value"));
             Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, null, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topScoreDocCollector.topDocs().totalHits.value);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                null,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topScoreDocAdapter.getResult().totalHits.value);
         }
     }
 
     public void testPostFilterWithAggs() throws IOException {
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
             TermQuery termQuery = new TermQuery(new Term("field1", "value"));
             Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, aggsCollector, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(numDocs, aggsCollector.getTotalHits());
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(numDocs, aggsAdapter.getResult().intValue());
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
             TermQuery termQuery = new TermQuery(new Term("field2", "value"));
             Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, aggsCollector, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
             // post_filter is not applied to aggs
-            assertEquals(reader.maxDoc(), aggsCollector.getTotalHits());
+            assertEquals(reader.maxDoc(), aggsAdapter.getResult().intValue());
         }
     }
 
@@ -196,33 +279,51 @@ public class QueryPhaseCollectorTests extends ESTestCase {
             .add(new BoostQuery(new TermQuery(new Term("field2", "value")), 200f), BooleanClause.Occur.SHOULD)
             .build();
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField2Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField2Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
             thresholdScore = topDocs.scoreDocs[numField2Docs].score;
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, null, maxScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(topScoreDocAdapter, null, 0, null, maxScore);
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, null, thresholdScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                null,
+                0,
+                null,
+                thresholdScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topScoreDocAdapter.getResult().totalHits.value);
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, null, maxScore + 100f);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(0, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                null,
+                0,
+                null,
+                maxScore + 100f
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(0, topScoreDocAdapter.getResult().totalHits.value);
         }
     }
 
@@ -234,46 +335,67 @@ public class QueryPhaseCollectorTests extends ESTestCase {
             .add(new BoostQuery(new TermQuery(new Term("field2", "value")), 200f), BooleanClause.Occur.SHOULD)
             .build();
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField2Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField2Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
             thresholdScore = topDocs.scoreDocs[numField2Docs].score;
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, aggsCollector, maxScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
-            // min_score is applied to aggs as well as top docs
-            assertEquals(numField2Docs, aggsCollector.getTotalHits());
-        }
-        {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, 0, aggsCollector, thresholdScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(numDocs, aggsCollector.getTotalHits());
-        }
-        {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(
-                topScoreDocCollector,
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
                 null,
                 0,
-                aggsCollector,
+                aggsAdapter,
+                maxScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
+            // min_score is applied to aggs as well as top docs
+            assertEquals(numField2Docs, aggsAdapter.getResult().intValue());
+        }
+        {
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                null,
+                0,
+                aggsAdapter,
+                thresholdScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(numDocs, aggsAdapter.getResult().intValue());
+        }
+        {
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                null,
+                0,
+                aggsAdapter,
                 maxScore + 100f
             );
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(0, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(0, aggsCollector.getTotalHits());
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(0, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(0, aggsAdapter.getResult().intValue());
         }
     }
 
@@ -287,33 +409,57 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         TermQuery termQuery = new TermQuery(new Term("field2", "value"));
         Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField3Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField3Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
             thresholdScore = topDocs.scoreDocs[numField3Docs].score;
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, null, maxScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2AndField3Docs, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                null,
+                maxScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2AndField3Docs, topScoreDocAdapter.getResult().totalHits.value);
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, null, thresholdScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                null,
+                thresholdScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, null, maxScore + 100f);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(0, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                null,
+                maxScore + 100f
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(0, topScoreDocAdapter.getResult().totalHits.value);
         }
     }
 
@@ -327,91 +473,130 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         TermQuery termQuery = new TermQuery(new Term("field2", "value"));
         Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField3Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField3Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
             thresholdScore = topDocs.scoreDocs[numField3Docs].score;
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, filterWeight, 0, aggs, maxScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2AndField3Docs, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(numField3Docs, aggs.getTotalHits());
-        }
-        {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(
-                topScoreDocCollector,
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
                 filterWeight,
                 0,
-                aggsCollector,
+                aggsAdapter,
+                maxScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2AndField3Docs, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(numField3Docs, aggsAdapter.getResult().intValue());
+        }
+        {
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
+                filterWeight,
+                0,
+                aggsAdapter,
                 thresholdScore
             );
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(numDocs, aggsCollector.getTotalHits());
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(numDocs, aggsAdapter.getResult().intValue());
         }
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggsCollector = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(
-                topScoreDocCollector,
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topScoreDocAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topScoreDocAdapter,
                 filterWeight,
                 0,
-                aggsCollector,
+                aggsAdapter,
                 maxScore + 100f
             );
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(0, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(0, aggsCollector.getTotalHits());
+            searcher.search(booleanQuery, manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(0, topScoreDocAdapter.getResult().totalHits.value);
+            assertEquals(0, aggsAdapter.getResult().intValue());
         }
     }
 
     public void testTerminateAfterTopDocsOnly() throws IOException {
         {
             int terminateAfter = randomIntBetween(1, numDocs - 1);
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, null, terminateAfter, null, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topDocs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                null,
+                terminateAfter,
+                null,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().intValue());
         }
         {
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, null, numDocs, null, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topDocs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(topDocsAdapter, null, numDocs, null, null);
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topDocsAdapter.getResult().intValue());
         }
     }
 
     public void testTerminateAfterWithAggs() throws IOException {
         {
             int terminateAfter = randomIntBetween(1, numDocs - 1);
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            DummyTotalHitCountCollector aggs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, null, terminateAfter, aggs, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topDocs.getTotalHits());
-            assertEquals(terminateAfter, aggs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                null,
+                terminateAfter,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().intValue());
+            assertEquals(terminateAfter, aggsAdapter.getResult().intValue());
         }
         {
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            DummyTotalHitCountCollector aggs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, null, numDocs, aggs, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numDocs, topDocs.getTotalHits());
-            assertEquals(numDocs, aggs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                null,
+                numDocs,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numDocs, topDocsAdapter.getResult().intValue());
+            assertEquals(numDocs, aggsAdapter.getResult().intValue());
         }
     }
 
@@ -420,19 +605,33 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
         {
             int terminateAfter = randomIntBetween(1, numField2Docs - 1);
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, filterWeight, terminateAfter, null, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topDocs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                filterWeight,
+                terminateAfter,
+                null,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().intValue());
         }
         {
             int terminateAfter = randomIntBetween(numField2Docs, Integer.MAX_VALUE);
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, filterWeight, terminateAfter, null, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topDocs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                filterWeight,
+                terminateAfter,
+                null,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topDocsAdapter.getResult().intValue());
         }
     }
 
@@ -441,25 +640,41 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
         {
             int terminateAfter = randomIntBetween(1, numField2Docs - 1);
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            DummyTotalHitCountCollector aggs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, filterWeight, terminateAfter, aggs, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topDocs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                filterWeight,
+                terminateAfter,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().intValue());
             // aggs see more docs because they are not filtered
-            assertThat(aggs.getTotalHits(), Matchers.greaterThanOrEqualTo(terminateAfter));
+            assertThat(aggsAdapter.getResult(), Matchers.greaterThanOrEqualTo(terminateAfter));
         }
         {
             int terminateAfter = randomIntBetween(numField2Docs, Integer.MAX_VALUE);
-            DummyTotalHitCountCollector topDocs = new DummyTotalHitCountCollector();
-            DummyTotalHitCountCollector aggs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topDocs, filterWeight, terminateAfter, aggs, null);
-            searcher.search(new MatchAllDocsQuery(), queryPhaseCollector);
-            assertFalse(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(numField2Docs, topDocs.getTotalHits());
+            CollectorManager<DummyTotalHitCountCollector, Integer> topDocsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                filterWeight,
+                terminateAfter,
+                aggsAdapter,
+                null
+            );
+            searcher.search(new MatchAllDocsQuery(), manager);
+            assertFalse(manager.isTerminatedEarly());
+            assertEquals(numField2Docs, topDocsAdapter.getResult().intValue());
             // aggs see more docs because they are not filtered
-            assertThat(aggs.getTotalHits(), Matchers.greaterThanOrEqualTo(numField2Docs));
+            assertThat(aggsAdapter.getResult(), Matchers.greaterThanOrEqualTo(numField2Docs));
         }
     }
 
@@ -470,19 +685,29 @@ public class QueryPhaseCollectorTests extends ESTestCase {
             .add(new BoostQuery(new TermQuery(new Term("field2", "value")), 200f), BooleanClause.Occur.SHOULD)
             .build();
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField2Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField2Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
         }
         {
             int terminateAfter = randomIntBetween(1, numField2Docs - 1);
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, terminateAfter, null, maxScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topScoreDocCollector.topDocs().totalHits.value);
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                null,
+                terminateAfter,
+                null,
+                maxScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().totalHits.value);
         }
     }
 
@@ -493,21 +718,32 @@ public class QueryPhaseCollectorTests extends ESTestCase {
             .add(new BoostQuery(new TermQuery(new Term("field2", "value")), 200f), BooleanClause.Occur.SHOULD)
             .build();
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField2Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField2Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
         }
         {
             int terminateAfter = randomIntBetween(1, numField2Docs - 1);
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(topScoreDocCollector, null, terminateAfter, aggs, maxScore);
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topScoreDocCollector.topDocs().totalHits.value);
-            assertEquals(terminateAfter, aggs.getTotalHits());
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
+                null,
+                terminateAfter,
+                aggsAdapter,
+                maxScore
+            );
+            searcher.search(booleanQuery, manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().totalHits.value);
+            assertEquals(terminateAfter, aggsAdapter.getResult().intValue());
         }
     }
 
@@ -520,25 +756,29 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         TermQuery termQuery = new TermQuery(new Term("field2", "value"));
         Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField3Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField3Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
         }
         {
             int terminateAfter = randomIntBetween(1, numField2AndField3Docs - 1);
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(
-                topScoreDocCollector,
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
                 filterWeight,
                 terminateAfter,
                 null,
                 maxScore
             );
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topScoreDocCollector.topDocs().totalHits.value);
+            searcher.search(booleanQuery, manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().totalHits.value);
         }
     }
 
@@ -551,28 +791,33 @@ public class QueryPhaseCollectorTests extends ESTestCase {
         TermQuery termQuery = new TermQuery(new Term("field2", "value"));
         Weight filterWeight = termQuery.createWeight(searcher, ScoreMode.TOP_DOCS, 1f);
         {
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(numField3Docs + 1, 1000);
-            searcher.search(booleanQuery, topScoreDocCollector);
-            TopDocs topDocs = topScoreDocCollector.topDocs();
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(
+                numField3Docs + 1,
+                null,
+                1000
+            );
+            TopDocs topDocs = searcher.search(booleanQuery, topDocsManager);
             assertEquals(numDocs, topDocs.totalHits.value);
             maxScore = topDocs.scoreDocs[0].score;
         }
         {
             int terminateAfter = randomIntBetween(1, numField2AndField3Docs - 1);
-            TopScoreDocCollector topScoreDocCollector = TopScoreDocCollector.create(1, 1000);
-            DummyTotalHitCountCollector aggs = new DummyTotalHitCountCollector();
-            QueryPhaseCollector queryPhaseCollector = new QueryPhaseCollector(
-                topScoreDocCollector,
+            CollectorManager<TopScoreDocCollector, TopDocs> topDocsManager = TopScoreDocCollector.createSharedManager(1, null, 1000);
+            CollectorManagerAdapter<TopScoreDocCollector, TopDocs> topDocsAdapter = new CollectorManagerAdapter<>(topDocsManager);
+            CollectorManager<DummyTotalHitCountCollector, Integer> aggsManager = DummyTotalHitCountCollector.createManager();
+            CollectorManagerAdapter<DummyTotalHitCountCollector, Integer> aggsAdapter = new CollectorManagerAdapter<>(aggsManager);
+            QueryPhaseCollector.CollectorManager manager = QueryPhaseCollector.createManager(
+                topDocsAdapter,
                 filterWeight,
                 terminateAfter,
-                aggs,
+                aggsAdapter,
                 maxScore
             );
-            searcher.search(booleanQuery, queryPhaseCollector);
-            assertTrue(queryPhaseCollector.isTerminatedAfter());
-            assertEquals(terminateAfter, topScoreDocCollector.topDocs().totalHits.value);
+            searcher.search(booleanQuery, manager);
+            assertTrue(manager.isTerminatedEarly());
+            assertEquals(terminateAfter, topDocsAdapter.getResult().totalHits.value);
             // aggs see more documents because the filter is not applied to them
-            assertThat(aggs.getTotalHits(), Matchers.greaterThanOrEqualTo(terminateAfter));
+            assertThat(aggsAdapter.getResult(), Matchers.greaterThanOrEqualTo(terminateAfter));
         }
     }
